rework the lexer to keep track of source locations for each token
In order to have the original CREATE TABLE text from the schema
available when using the parsed structure, rework the lexer so that
instead of extracting the substring tokens, it extracts tuples of
[position, length] and maintains these associations when collapsing
the token sequences.

This means we can include the original source text in a `sql` field
of the resulting parsed structure.